fixes #30 #8
