+++ src/ui/AppProjectInfo.tsx
@@ -1,10 +1,12 @@
-//TODO: find a better name for this class
-
 import { List, ListItem, ListItemIcon, ListItemText } from "@material-ui/core"
 import CodeIcon from '@material-ui/icons/Code';
 import { version } from '../../package.json'
 import { GitHubMaterialIcon } from "."
 
+/**
+ * A List containing all relevant App Informations e.g. put in drawer
+ * @returns JSX.Element
+ */
 export const AppProjectInfo = () => {
     return (
         <List>
