+++ src/ui/GithubIcon.tsx
@@ -3,6 +3,11 @@ import Icon from '@iconify/react';
 
 
 //both icons named Icon I need to split
+
+/**
+ * Apstratction to create a GitHub Icon 
+ * @returns Github Icon
+ */
 export const GitHubIcon = () => {
     return(
             <Icon icon={githubIcon} />
