+++ src/ui/dialog/GeneralDescriptionDialog.tsx
@@ -1,10 +1,20 @@
 import AppDialog from "./AppDialog"
 
 interface Props{
+    /**
+     * On close clicked
+     */
     buttonCloseClicked:()=> void
+    /**
+     * state value if is displayed
+     */
     isOpen:boolean
 }
-
+/**
+ * Displays the General game description extends AppDialog
+ * @param props Props
+ * @returns A Dialog
+ */
 export const GeneralDescriptionDialog = (props:Props) => {
     return <AppDialog 
     title={"Welcome"}
