+++ src/ui/index.ts
@@ -5,3 +5,4 @@ export * from "./AppProjectInfo"
 export * from "./DisplayLoading"
 export type {DialogType} from "./ContentMainDialogeFactory"
 export * from "./dialog/UnsupportedBrowserSupport"
+export * from "./MainDrawer"
